Support `postman` and `har` inputs for capture 2.0 (#2121)
